+++ src/utils/mod.rs
@@ -3,6 +3,7 @@ pub mod color;
 pub mod file_scanner;
 pub mod filters;
 pub mod float_map;
+pub mod frame_cache;
 pub mod image_loader;
 pub mod lc_presets;
 pub mod local_contrast;